requirements.txt
bitcoinlib==0.6.12
bitcoinrpc==0.6.1
Flask==3.0.0
matplotlib==3.7.2
model==0.6.0
numpy==1.24.4
opencv_python_headless==4.8.1.78
pandas==2.1.2
Pillow==10.0.1
Pillow==7.2.0
Pillow==10.1.0
plotly==5.17.0
Requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
torch==2.1.0
torchvision==0.15.2a0
transformers==4.33.2
